requirements.txt
requests==2.25.0
beautifulsoup4==4.9.3
dnspython==2.0.0
python_Levenshtein==0.12.0
